consolidate two identical function definitions
of get_cases_in_domain
