+++ corehq/apps/hqcase/utils.py
@@ -1,22 +1,21 @@
 import datetime
 import uuid
 from xml.etree import ElementTree
+import xml.etree.ElementTree as ET
+import re
+
 from couchdbkit import ResourceNotFound
 from django.core.files.uploadedfile import UploadedFile
+from django.template.loader import render_to_string
+
 from casexml.apps.phone.xml import get_case_xml
-from corehq.apps.hqcase.dbaccessors import get_case_ids_in_domain
-from dimagi.utils.couch.database import iter_docs
 from casexml.apps.case.mock import CaseBlock
 from casexml.apps.case.models import CommCareCase
 from dimagi.utils.parsing import json_format_datetime
-from django.template.loader import render_to_string
 from casexml.apps.case.xml import V2
 from casexml.apps.phone.caselogic import get_related_cases
 from corehq.apps.hqcase.exceptions import CaseAssignmentError
-
 from corehq.apps.receiverwrapper import submit_form_locally
-import xml.etree.ElementTree as ET
-import re
 from casexml.apps.case import const
 
 
@@ -136,11 +135,6 @@ def get_case_by_identifier(domain, identifier):
         pass
 
     return None
-
-
-def get_cases_in_domain(domain, type=None):
-    return (CommCareCase.wrap(doc) for doc in iter_docs(CommCareCase.get_db(),
-                                                        get_case_ids_in_domain(domain, type=type)))
 
 
 def assign_case(case_or_case_id, owner_id, acting_user=None, include_subcases=True,
