+++ custom/bihar/management/commands/bihar_create_backlogged_repeaters.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 import logging
 from django.core.management.base import BaseCommand
-from corehq.apps.hqcase.utils import get_cases_in_domain
+from corehq.apps.hqcase.dbaccessors import get_cases_in_domain
 from corehq.apps.receiverwrapper.models import CaseRepeater
